+++ server.js
@@ -6,12 +6,6 @@ var async = require('async');
 var nopt = require('nopt');
 
 var openVeoApi = require('@openveo/api');
-var ClientProvider = process.require('app/server/providers/ClientProvider.js');
-var RoleProvider = process.require('app/server/providers/RoleProvider.js');
-var TokenProvider = process.require('app/server/providers/TokenProvider.js');
-var UserProvider = process.require('app/server/providers/UserProvider.js');
-var GroupProvider = process.require('app/server/providers/GroupProvider.js');
-var TaxonomyProvider = process.require('app/server/providers/TaxonomyProvider.js');
 var CorePlugin = process.require('app/server/plugin/CorePlugin.js');
 var storage = process.require('app/server/storage.js');
 var api = process.require('app/server/api.js');
@@ -192,32 +186,6 @@ async.series([
   function(callback) {
     var migrations = server.migrations;
     migrationProcess.executeMigrationScript(migrations, callback);
-  },
-
-  // Create core indexes
-  function(callback) {
-    var database = storage.getDatabase();
-    var asyncFunctions = [];
-    var providers = [
-      new ClientProvider(database),
-      new RoleProvider(database),
-      new TaxonomyProvider(database),
-      new TokenProvider(database),
-      new UserProvider(database),
-      new GroupProvider(database)
-    ];
-
-    providers.forEach(function(provider) {
-      if (provider.createIndexes) {
-        asyncFunctions.push(function(callback) {
-          provider.createIndexes(callback);
-        });
-      }
-    });
-
-    async.parallel(asyncFunctions, function(error, results) {
-      callback(error);
-    });
   },
 
   // Intitializes plugins