+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## BUG FIXES
 
 - Fix back end menu items positions. Items without sub items were organized between them and items with sub items were organized between them. There is now no distinction made between items without sub items and items with sub items
+- Automatically remove role from users when a role is deleted
 
 # 4.1.0 / 2017-09-12